third_party/WebKit/LayoutTests/media/media-constants.html

<!DOCTYPE html>
<title>Test HTMLMediaElement and MediaError constants.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
test(function() {
    assert_equals(HTMLMediaElement.NETWORK_EMPTY, 0);
    assert_equals(HTMLMediaElement.NETWORK_IDLE, 1);
    assert_equals(HTMLMediaElement.NETWORK_LOADING, 2);
    assert_equals(HTMLMediaElement.NETWORK_NO_SOURCE, 3);

    assert_equals(HTMLMediaElement.HAVE_NOTHING, 0);
    assert_equals(HTMLMediaElement.HAVE_METADATA, 1);
    assert_equals(HTMLMediaElement.HAVE_CURRENT_DATA, 2);
    assert_equals(HTMLMediaElement.HAVE_FUTURE_DATA, 3);
    assert_equals(HTMLMediaElement.HAVE_ENOUGH_DATA, 4);

    assert_equals(MediaError.MEDIA_ERR_ABORTED, 1);
    assert_equals(MediaError.MEDIA_ERR_NETWORK, 2);
    assert_equals(MediaError.MEDIA_ERR_DECODE, 3);
    assert_equals(MediaError.MEDIA_ERR_SRC_NOT_SUPPORTED, 4);
});
</script>
